move SizeInput from lib into chunky ui package
