+++ chunky/src/res/se/llbit/chunky/ui/render/tabs/GeneralTab.fxml
@@ -4,6 +4,7 @@
 <?import javafx.scene.control.*?>
 <?import javafx.scene.layout.*?>
 <?import se.llbit.chunky.ui.*?>
+<?import se.llbit.chunky.ui.elements.*?>
 
 <fx:root type="javafx.scene.control.ScrollPane" xmlns="http://javafx.com/javafx/8.0.40" xmlns:fx="http://javafx.com/fxml/1">
   <VBox spacing="10.0">
